% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basic_stats_functions.R
\name{agencies_populous_title_stats}
\alias{agencies_populous_title_stats}
\title{Agency most populous titles and their composition
TASKS that use this:
Agency Findings}
\usage{
agencies_populous_title_stats(data, top_n_titles = 1)
}
\arguments{
\item{data}{Pass the full or subset of the cleaned dataset using \code{payeqfunctions::clean_data()} to this function}

\item{top_n_titles}{Number of top populous titles per agency. Default is set to 1}
}
\value{
Table of a variable input's breakdown in
}
\description{
NOTE: This function will ONLY work if the data input is generated by \code{payeqfunctions::clean_data()} due to the naming scheme.
}
\examples{
\dontrun{
# Agency Findings
agencies_populous_title_stats(data = cleaned_dataset)
}
}
