% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_data}
\alias{clean_data}
\title{Cleaning the dataset}
\usage{
clean_data(
  filepath = "C:/Users/Public/Desktop/1- 22 Data Elements Dataset.csv",
  year,
  column_names = "default",
  filter_employee_status = "full-time",
  uniform_filepath
)
}
\arguments{
\item{filepath}{Path to the pay equity file. Default is set as "C:/Users/Public/Desktop/1- 22 Data Elements Dataset.csv"}

\item{year}{YYYY Year of dataset. Used for calculating years from start and age}

\item{column_names}{Accepts a vector of column names in the order they should appear. Default is \code{"default"}. WARNING: If the default does not work, or is changed, other functions in this package may not function}

\item{filter_employee_status}{Filter options, default is \code{filter = "full-time"}. Accepts \code{"part-time"} for filtering to a part-time prefixed settings. If anything else is inputted, then it will return the full dataset without any employee_status filters.}

\item{uniform_filepath}{Add filepath for the uniform titles for the uniform title column}
}
\value{
Cleaned dataset as a data frame
}
\description{
This function will ask for the file location containing the pay equity data.
It will rename all the column names, then clean and set data types the correct ones.
The newly created columns are as follows:
"days_from_start":
"years_from_start":
"age":
"age_years":
"age_above18":
"race_eth":
"nonwhite":
"nonwhite_female":
}
\details{
This function will return a cleaned and tidied pay dataset.

By default, dataset is filtered for "Full-Time", "Competitive", and "Non-Competitive" employees who make over $15,000 and are over 16 years of age.
}
