% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basic_stats_functions.R
\name{gender_summary}
\alias{gender_summary}
\title{Gender composition by agency title names and title codes}
\usage{
gender_summary(data, interested_agency, exportable = FALSE)
}
\arguments{
\item{data}{dataframe. Pass the full or subset of the cleaned dataset using \code{payeqfunctions::clean_data()} to this}

\item{interested_agency}{string. Full name of the agency as it shows in the dataset}

\item{exportable}{boolean. Option to generate a table excluding all rows with < 5 gender in the title. Default is \code{FALSE}}
}
\value{
Dataframe of gender composition per title at the specified agency
}
\description{
The output is a table of all the civil service titles and civil service title codes in a specified agency, and their corresponding gender makeup.
}
\details{
TASKS that use this:
\itemize{
\item FDNY Salaries by Gender Composition
\item NYPD Salaries by Gender Composition
\item DOITT Salaries by Gender Composition
\item DSNY Salaries by Gender Composition
}

NOTE: This function will ONLY work if the data input is generated by \code{payeqfunctions::clean_data()} due to the naming scheme.
}
\examples{
\dontrun{
# FDNY Salaries by Gender Composition
gender_summary(data = cleaned_dataset, interested_agency = "FIRE DEPARTMENT")
# DSNY Salaries by Gender Composition w/ exportable option
gender_summary(data = cleaned_dataset, interested_agency = "DEPARTMENT OF SANITATION", exportable = TRUE)
}
}
