% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_code.R
\name{log_to_prop}
\alias{log_to_prop}
\title{Log coefficent conversion table helper function}
\usage{
log_to_prop(gender_val, race_eth_val, coef_df)
}
\arguments{
\item{gender_val}{Name of the gender coef variable}

\item{race_eth_val}{Name of the race_eth coef variable}

\item{coef_df}{The coef data frame generated from the model}
}
\value{
Table of pay gap and adjusted pay gap based on model coefficients
}
\description{
Log coefficent conversion table helper function
}
