% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_code.R
\name{pay_gap_table}
\alias{pay_gap_table}
\title{Base Hierarchical and Interaction Model}
\usage{
pay_gap_table(data, interaction_model = FALSE)
}
\arguments{
\item{data}{Pass the cleaned dataset using \code{payeqfunctions::clean_data()} to this}

\item{interaction_model}{Default is FALSE. When TRUE, it'll run the interaction model}
}
\value{
List of pay gap tables. When running the non-interaction model, the first item on the list is the gender pay table and second item is race_eth pay table
}
\description{
This function will create either a \code{basic_model} or \code{interaction_model} in your global environment. Additionally, it will create a paygap table based on the coefficients
}
\details{
TASKS that use this:
\itemize{
\item Gender (unadjusted too)
\item Race/Ethnicity (unadjusted too)
\item Gender & Race/Ethnicity (unadjusted too)
}

NOTE: This function will ONLY work if the data input is generated by \code{payeqfunctions::clean_data()} due to the naming scheme.
}
\examples{
\dontrun{
# Regular lme model
pay_gap_table(data = cleaned_dataset)
# output[[1]] for gender pay gap table
# output[[2]] for race_eth pay gap table

# Interaction lme model
pay_gap_table(data = cleaned_dataset, interaction_model = TRUE)
}
}
