% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basic_stats_functions.R
\name{race_eth_summary}
\alias{race_eth_summary}
\title{Race / Ethnicity composition}
\usage{
race_eth_summary(
  data,
  interested_agency = NULL,
  high_low = NULL,
  spread_format = "wide",
  exportable = FALSE
)
}
\arguments{
\item{data}{Pass the full or subset of the cleaned dataset using \code{payeqfunctions::clean_data()} to this}

\item{interested_agency}{Full name of the agency as it shows in the dataset}

\item{high_low}{Option for the highest paid or lowest paid titles. Default is \code{"NULL"}, and accepts \code{"high"} and \code{"low"} for filtering to the highest and lowest paid titles respectively}

\item{spread_format}{Option to have the output wide or long. Default is \code{"wide"}, and accepts \code{"long"} for wide or long data respectively}
}
\value{
Table of race / ethnicity composition per title at the specified agency
}
\description{
This is giant compilation of code that does a race_eth breakdown.
NOTE: This function will ONLY work if the data input is generated by \code{payeqfunctions::clean_data()} due to the naming scheme.
}
\details{
TASKS that use this:
Race/Ethnicity Breakdown for Lowest-Paying Civil Service Titles
Race/Ethnicity Breakdown for Highest-Paying Civil Service Titles
}
\examples{
\dontrun{
# Race/Ethnicity Breakdown for Lowest-Paying Civil Service Titles
race_eth_summary(data = cleaned_dataset, high_low = "low", spread_format = "long")
# Race/Ethnicity Breakdown for Highest-Paying Civil Service Titles
race_eth_summary(data = cleaned_dataset, high_low = "high", spread_format = "long")
# Race/Ethnicity Breakdown by Title in an Agency
race_eth_summary(data = cleaned_dataset, interested_agency = "FIRE DEPARTMENT")
}
}
