% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basic_stats_functions.R
\name{title_code_share}
\alias{title_code_share}
\title{Proportion of a variable of interest in titles}
\usage{
title_code_share(data, share_var, share_var_val, binning = TRUE)
}
\arguments{
\item{data}{dataframe. Pass the full or subset of the cleaned dataset using \code{payeqfunctions::clean_data()} to this function}

\item{share_var}{string. Variable name that is being filtered.}

\item{share_var_val}{string. The value the filtered variable should take.}

\item{binning}{Default is TRUE. When TRUE, returns a binned (with a binwidth of 0.1) table. When FALSE, returns a dataset of all the titles in the passed data with the percentage of the variable of interest}
}
\value{
Dataframe of titles with details or a binned version.
}
\description{
This gives a summarised table of titles with their percentage of the variable of interest. Binning is another layer of grouping available.
}
\details{
TASKS that use this:
\itemize{
\item Median Salary for Civil Service Title Codes by Share of Non-White Employees per Title
\item Median Salary for Civil Service Title Codes by Share of Female Employees per Title
\item Median Salary for Civil Service Title Codes by Share of Non-White Female Employees per Title
}

NOTE: This function will ONLY work if the data input is generated by \code{payeqfunctions::clean_data()} due to the naming scheme.
}
\examples{
\dontrun{
# Median Salary for Civil Service Title Codes by Share of Non-White Employees per Title
title_code_share(data = cleaned_dataset, share_var = "nonwhite", share_var_val = 1) # Binned nonwhite proportions for titles
# Median Salary for Civil Service Titltle Codes by Share of Female Employees per Title
title_code_share(data = cleaned_dataset, share_var = "gender", share_var_val = "Female")
# Median Salary for Civil Service Title Codes by Share of Non-White Female Employees per Title
title_code_share(data = DOB_cleaned_dataset, share_var = "nonwhite_female", share_var_val = 1, binning = FALSE) # Nonbinned nonwhite female proportions for titles in the DOB
}
}
