% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{view_default_column_names}
\alias{view_default_column_names}
\title{Viewing column names used for clean_data function}
\usage{
view_default_column_names()
}
\value{
Vector of column names used
}
\description{
Viewing column names used for clean_data function. Useful as a jumping board should any column names need to be changed.
}
\examples{
view_default_column_names()
}
